feat(Endpoint): Eliminate device handler inside endpoint
This commit removed the device handler inside the HAL directory,
since the specific device type could be decided during compile-time.

The HAL layer has been changed to stub. The common usage of function
has been defined inside "endpoint/platform/impl.h". The endpoint
simulator stub has been implemented inside
"endpoint/platform/simulator/impl.c". And the stub for legato targets
has been implemented inside "endpoint/platform/legato-target/impl.c".

Use the stub for legato target will avoid to define new
HAL target file again when a new type of target was added.

However, we only test these on fewer devices. Such as wp77xx and FX30.
Other legato device has not been ensured.

The UART stub for simulator and legato-target has been removed since
the legato framework provides UART API to configure serial ports.

Close #707
