+++ endpoint/OBDComp/endpointService/Component.cdef
@@ -1,10 +1,13 @@
 sources:
 {
     ${CURDIR}/../../endpoint_core.c
-    ${CURDIR}/../../hal/device.c
     
     // include the specific platform
-    ${CURDIR}/../../platform/${EP_TARGET}/impl.c 
+    #if ${LEGATO_TARGET} = localhost
+        ${CURDIR}/../../platform/simulator/impl.c
+    #else
+        ${CURDIR}/../../platform/legato-target/impl.c
+    #endif
 
     ${CURDIR}/../../../output_base/external/org_iota_common/utils/logger_helper.c
