--- endpoint/hal/BUILD
@@ -1,10 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "hal",
-    srcs = ["device.c"],
-    hdrs = ["device.h"],
-    deps = [
-        "//common:ta_errors",
-    ],
-)
