--- endpoint/hal/device.c
@@ -1,62 +1,0 @@
-/*
- * Copyright (C) 2019-2020 BiiLabs Co., Ltd. and Contributors
- * All Rights Reserved.
- * This is free software; you can redistribute it and/or modify it under the
- * terms of the MIT license. A copy of the license can be found in the file
- * "LICENSE" at the root of this distribution.
- */
-
-#include "legato.h"
-
-#include "device.h"
-#include "le_log.h"
-
-static struct device_type *devices;
-
-static struct device_type **find_device(const char *name, unsigned len) {
-  struct device_type **p;
-  for (p = &devices; *p; p = &(*p)->next)
-    if (strlen((*p)->name) == len && !strncmp((*p)->name, name, len)) break;
-  return p;
-}
-
-device_t *ta_device(const char *type) {
-  struct device_type **p;
-  if (devices->next) {
-    LE_ERROR("No device type registered!");
-    return NULL;
-  }
-  p = find_device(type, strlen(type));
-  if (*p == NULL) {
-    LE_ERROR("Device type %s not found", type);
-    return NULL;
-  }
-  LE_DEBUG("Successfully get %s device", type);
-  return *p;
-}
-
-status_t register_device(struct device_type *dv) {
-  status_t res = SC_OK;
-  struct device_type **p;
-  if (dv->next) {
-    return SC_ENDPOINT_DEVICE_INIT;
-  }
-  p = find_device(dv->name, strlen(dv->name));
-  if (*p) {
-    res = SC_ENDPOINT_DEVICE_INIT;
-  } else {
-    *p = dv;
-  }
-  return res;
-}
-
-status_t unregister_device(struct device_type *dv) {
-  for (struct device_type **tmp = &devices; *tmp != NULL; tmp = &(*tmp)->next) {
-    if (dv == *tmp) {
-      *tmp = dv->next;
-      dv->next = NULL;
-      return SC_OK;
-    }
-  }
-  return SC_ENDPOINT_DEVICE_FINI;
-}
