--- endpoint/platform/BUILD
@@ -1,9 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "device",
-    deps = [
-        "//endpoint/hal",
-        "//endpoint/simluator",
-    ],
-)
