--- endpoint/platform/simulator/BUILD
@@ -1,9 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "simulator",
-    srcs = ["simulator.c"],
-    deps = [
-        "//endpoint/hal",
-    ],
-)
