--- endpoint/platform/wp77xx/BUILD
@@ -1,9 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "wp77xx",
-    srcs = ["wp77xx.c"],
-    deps = [
-        "//endpoint/hal",
-    ],
-)
